requirements.txt
# AI / Agent frameworks
anthropic>=0.25.0
google-adk>=0.1.0
google-generativeai
google-genai

# Web framework
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9

# Environment & config
python-dotenv>=1.0.0

# Analysis & visualization
jsbeautifier>=1.14.0
networkx
matplotlib
scipy
cryptography
requests

# Sandbox execution
e2b>=2.0.0
e2b_code_interpreter
